requirement.txt
anyio==3.6.1
bcrypt==4.0.0
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
fastapi==0.85.0
future==0.18.2
h11==0.13.0
idna==3.4
mslib==0.1.2
mysql-connector==2.2.9
netmiko==4.1.2
ntc-templates==3.1.0
numpy==1.23.5
pandas==1.5.2
paramiko==2.11.0
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
PyNaCl==1.5.0
pyserial==3.5
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.6
PyYAML==6.0
scp==0.14.4
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
tenacity==8.0.1
textfsm==1.1.2
typing_extensions==4.3.0
uvicorn==0.18.3
